Add files via upload
This is a basic quiz game built in python version 3.11.5. The main library used here is tkinter.  The quiz game starts with a welcome window for the user where the user will have to insert his/her name and click on submit. Later on the main quiz interface will appear which is user friendly and easy to use. On the left side of the interface there are total number of questions and the highlighted one is the current ongoing question. The button options change color when the cursor is hovered over it. If user is unable to solve a question then the user can skip the question using skip button in the bottom.

Once the quiz is done the results will be displayed and star badges will be assigned based on you success rate. Lastly there will be a thanking note for the participation. In the end, the user can click on "Go to home" button to get back on the welcoming window.